backend/pyproject.toml
[project]
name = "KitchenOwl-Backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apispec>=6.8.1",
    "apscheduler>=3.11.0",
    "blurhash-python>=1.2.2",
    "flask>=3.1.0",
    "flask-apscheduler>=1.13.1",
    "flask-basicauth>=0.2.0",
    "flask-bcrypt>=1.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-migrate>=4.1.0",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "gevent>=24.11.1",
    "greenlet>=3.1.1",
    "ingredient-parser-nlp>=2.0.0",
    "litellm>=1.65.0",
    "marshmallow>=3.26.1",
    "mlxtend==0.23.1",
    "numpy>=2.2.4",
    "prometheus-flask-exporter>=0.23.2",
    "oic>=1.7.0",
    "recipe-scrapers>=15.6.0",
    "pyuwsgi>=2.0.28",
    "celery>=5.4.0",
    "sqlite-icu>=1.0",
    "lark>=1.2.2",
    "dbscan1d>=0.2.3",
    "pyjwt==2.9.0",
    "pandas>=2.2.3",
    "prometheus-client>=0.21.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "tqdm>=4.67.1",
    "werkzeug>=3.1.3",
    "nltk>=3.9.1",
]

[dependency-groups]
dev = [
    "pytest", 
    "pytest-cov",
    "ruff",
    "pyright"
]

[tool.pyright]
reportMissingTypeStubs = true
typeCheckingMode = "strict"